ui/backend/requirements.txt
flask
flask-cors
numpy
pillow
tensorflow
mysql-connector-python
flask-bcrypt
pyjwt
python-dotenv